Components/App: Local company packs cache
